refactor(provider/kubernetes): Move ops w/ kind into v1 only
