+++ clouddriver-kubernetes/src/main/groovy/com/netflix/spinnaker/clouddriver/kubernetes/v1/deploy/description/securitygroup/KubernetesSecurityGroupDescription.groovy
@@ -16,13 +16,13 @@
 
 package com.netflix.spinnaker.clouddriver.kubernetes.v1.deploy.description.securitygroup
 
-import com.netflix.spinnaker.clouddriver.kubernetes.v1.deploy.description.KubernetesAtomicOperationDescription
+import com.netflix.spinnaker.clouddriver.kubernetes.v1.deploy.description.KubernetesKindAtomicOperationDescription
 import groovy.transform.AutoClone
 import groovy.transform.Canonical
 
 @AutoClone
 @Canonical
-class KubernetesSecurityGroupDescription extends KubernetesAtomicOperationDescription {
+class KubernetesSecurityGroupDescription extends KubernetesKindAtomicOperationDescription {
   String securityGroupName
   String app
   String stack
